Replace jackson for gson

+++ src/main/java/io/github/zeroone3010/yahueapi/DimmerSwitchImpl.java
@@ -23,6 +23,6 @@ final class DimmerSwitchImpl extends BasicSensor implements DimmerSwitch {
 
   @Override
   public DimmerSwitchButtonEvent getLatestButtonEvent() {
-    return new DimmerSwitchButtonEvent(readStateValue("buttonevent", Integer.class));
+    return new DimmerSwitchButtonEvent((int) (double) readStateValue("buttonevent", Double.class));
   }
 }
